refactor: LottoNumberGenerator를 util로 변경
