+++ src/main/java/lotto/domain/LottoMachine.java
@@ -4,21 +4,19 @@ import static lotto.constant.LottoConstant.LOTTO_PRICE;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.util.LottoNumberGenerator;
 
 public class LottoMachine {
     private final int lottoCount;
-    private final LottoNumberGenerator lottoNumberGenerator;
 
-    public LottoMachine(PurchaseAmount purchaseAmount,
-                        LottoNumberGenerator lottoNumberGenerator) {
+    public LottoMachine(PurchaseAmount purchaseAmount) {
         this.lottoCount = (int)(purchaseAmount.getAmount() / LOTTO_PRICE);
-        this.lottoNumberGenerator = lottoNumberGenerator;
     }
 
     public List<Lotto> issue() {
         List<Lotto> lottos = new ArrayList<>(lottoCount);
         for (int i = 0; i < lottoCount; i++) {
-            Lotto lotto = new Lotto(lottoNumberGenerator.generate());
+            Lotto lotto = new Lotto(LottoNumberGenerator.generate());
             lottos.add(lotto);
         }
         return List.copyOf(lottos);
